+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/ServiceDefinitionCollector.java
@@ -150,7 +150,8 @@ final class ServiceDefinitionCollector {
     private LoadId idOf(ExecutableElement x) {
         ServiceId annotation = x.getAnnotation(ServiceId.class);
         return new LoadId(x,
-                Optional.ofNullable(getServiceTypeOrNull(x))
+                Optional.ofNullable(getServiceTypeOrNull(x)),
+                annotation.pattern()
         );
     }